Rename `apiHost` option to `site`
This renames the `apiHost` option for `init()` and `BufferedMetricsLogger` to `site` in order to better align with Datadog docs and official packages. The old name is kept for now, but logs a deprecation warning to the console (regardless of the user's DEBUG setting) the first time it is used. It also adds tooling for logging similar deprecation messages and centralizes debug logging so we don't have to worry about typos when people define the debug logger in a module.